requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
google-generativeai==0.4.1
Pillow==10.2.0
diffusers==0.27.2
transformers==4.38.2
torch==2.2.1
accelerate==0.27.2
python-dotenv==1.0.1
